Do location calculations with doubles rather than ints. This simplifies the code and makes Zebra work.

+++ aisleriot/draw.h
@@ -25,7 +25,7 @@ extern int window_width;
 extern int window_height;
 
 void rescale_cards (void);
-void set_geometry (int new_width, int new_height);
+void set_geometry (double new_width, double new_height);
 void draw_cards( void );
 void take_snapshot( void );
 void refresh_screen( void );
